# docs/swagger.yaml
definitions:
  interfaces.StatusDetailDTO:
    properties:
      created_at:
        type: string
      status:
        type: string
      status_description:
        type: string
      task_id:
        type: integer
      updated_by:
        type: string
    type: object
  interfaces.StatusHistoryResponseDTO:
    properties:
      history:
        items:
          $ref: '#/definitions/interfaces.StatusDetailDTO'
        type: array
      message:
        type: string
      success:
        type: boolean
      total_count:
        type: integer
    type: object
  interfaces.TaskDetailDTO:
    properties:
      client_id:
        type: string
      client_name:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      position:
        type: string
    type: object
  interfaces.TasksResponseDTO:
    properties:
      message:
        type: string
      success:
        type: boolean
      tasks:
        items:
          $ref: '#/definitions/interfaces.TaskDetailDTO'
        type: array
      total_count:
        type: integer
    type: object
  schemas.TaskImportResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      imported_at:
        type: string
      message:
        type: string
      stats:
        $ref: '#/definitions/schemas.TaskImportStats'
      success:
        type: boolean
      total_entries:
        type: integer
    type: object
  schemas.TaskImportStats:
    properties:
      duration_ms:
        description: Changed from time.Duration to int64
        type: integer
      end_time:
        type: string
      error_count:
        type: integer
      start_time:
        type: string
      success_count:
        type: integer
      total_processed:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/commands/import:
    post:
      consumes:
      - application/json
      description: Triggers the task import process from a specified directory
      produces:
      - application/json
      responses:
        "200":
          description: Successful import response
          schema:
            $ref: '#/definitions/schemas.TaskImportResponse'
        "500":
          description: Error import response
          schema:
            $ref: '#/definitions/schemas.TaskImportResponse'
      summary: Import tasks from CSV
      tags:
      - commands
  /api/queries/tasks/active:
    get:
      description: Retrieves all active tasks for a specific client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interfaces.TasksResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/interfaces.TasksResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/interfaces.TasksResponseDTO'
      security:
      - Bearer: []
      summary: Get active tasks for a client
      tags:
      - queries
  /api/queries/tasks/history:
    get:
      description: Retrieves the status history of all tasks for a client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interfaces.StatusHistoryResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/interfaces.StatusHistoryResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/interfaces.StatusHistoryResponseDTO'
      security:
      - Bearer: []
      summary: Get task status history for a client
      tags:
      - queries
swagger: "2.0"